added dao and updated the code structure
