+++ endpoints/assets.py
@@ -1,18 +1,7 @@
 from flask import Blueprint, request, jsonify, send_from_directory
-from pathlib import Path
-import secrets, string
+from dao.asset_dao import save_file, get_file_path, UPLOAD_DIR
 
 assets_bp = Blueprint('assets', __name__)
-
-UPLOAD_DIR = Path("storage/images")
-UPLOAD_DIR.mkdir(parents=True, exist_ok=True)
-
-
-def generate_random_filename(original_filename):
-    ext = Path(original_filename).suffix 
-    random_name = ''.join(secrets.choice(string.ascii_letters) for i in range(10))  
-    return f"{random_name}{ext}"
-
 
 @assets_bp.route('/', methods=['POST'])
 def upload():
@@ -27,16 +16,15 @@ def upload():
     if not file.content_type.startswith('image/'):
         return jsonify({'error': 'File must be an image'}), 400
 
-    filename = generate_random_filename(file.filename)
-    file_path = UPLOAD_DIR / filename
-    file.save(file_path)
+    filename = save_file(file)
 
     return jsonify({'message': 'File uploaded successfully', 'filename': filename}), 200
 
-
 @assets_bp.route('/<filename>', methods=['GET'])
 def download(filename):
-    file_path = UPLOAD_DIR / filename
-    if not file_path.exists():
+    file_path = get_file_path(filename)
+    
+    if not file_path:
         return jsonify({'error': 'File not found'}), 404
+    
     return send_from_directory(UPLOAD_DIR, filename, as_attachment=True)
